requirements.txt
Flask
numpy
torch
torchvision
pandas
matplotlib
scikit-learn
tqdm
Pillow
